reorder the app
