+++ src/app/config.py
@@ -1,11 +1,12 @@
 # config of the apps
 
-import logging
 import os
 
 from dotenv import load_dotenv
 
-from src.lib.config import Config, init_logging as base_init_logging
+from src.lib.config import Config
+from src.lib.config import init_logging as base_init_logging
+
 
 class AppConfig(Config):
     APP_HOST: str = "localhost"
@@ -16,7 +17,7 @@ class AppConfig(Config):
     STORAGE_PRIVATE = "/private"
     STORAGE_DATA = "/data.json"
 
-    SINGLE_SEGMENT_SIZE_MIN = 10    # [byte]
+    SINGLE_SEGMENT_SIZE_MIN = 10  # [byte]
     SINGLE_SEGMENT_SIZE_MAX = 150  # [byte]
     WINDOW_TIMEOUT_MIN = 0.1  # [s]
     WINDOW_TIMEOUT_MAX = 1.0  # [s]
@@ -39,4 +40,3 @@ def init_config() -> None:
 
 def init_logging() -> None:
     base_init_logging()
-
